Clean up temporary test files reliably

+++ tests/Handler/MockHandlerTest.php
@@ -77,16 +77,21 @@ class MockHandlerTest extends TestCase
     public function testSinkFilename()
     {
         $filename = \sys_get_temp_dir().'/mock_test_'.\uniqid();
-        $res = new Response(200, [], 'TEST CONTENT');
-        $mock = new MockHandler([$res]);
-        $request = new Request('GET', '/');
-        $p = $mock($request, ['sink' => $filename]);
-        $p->wait();
 
-        self::assertFileExists($filename);
-        self::assertStringEqualsFile($filename, 'TEST CONTENT');
+        try {
+            $res = new Response(200, [], 'TEST CONTENT');
+            $mock = new MockHandler([$res]);
+            $request = new Request('GET', '/');
+            $p = $mock($request, ['sink' => $filename]);
+            $p->wait();
 
-        \unlink($filename);
+            self::assertFileExists($filename);
+            self::assertStringEqualsFile($filename, 'TEST CONTENT');
+        } finally {
+            if (\file_exists($filename)) {
+                \unlink($filename);
+            }
+        }
     }
 
     public function testSinkResource()
